requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dill==0.4.0
Flask==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
macholib==1.16.3
MarkupSafe==3.0.2
modulegraph==0.19.6
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
py2app==0.28.8
pycparser==2.22
requests==2.32.4
setuptools==80.9.0
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
